requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==2.3.0
python-jose==3.3.0
cryptography==41.0.3
python-dotenv==1.0.0
passlib==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pyopenssl==23.2.0
bcrypt==4.0.1 